# components/apimgt/org.wso2.carbon.apimgt.rest.api.perapilogger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This document specifies a **RESTful API** for WSO2 **API Manager** - Gateway",
    "version" : "v0.9",
    "title" : "WSO2 API Manager - Gateway",
    "contact" : {
      "name" : "WSO2",
      "url" : "http://wso2.com/products/api-manager/",
      "email" : "architecture@wso2.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath" : "/api/am/gateway/v0.9",
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/api-logging" : {
      "get" : {
        "parameters" : [ {
          "name" : "context",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "schema" : {
              "$ref" : "#/definitions/APIList"
            }
          },
          "404" : {
            "description" : "Not Found.\nRequest API resource or external store Ids not found.\n",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal server error while retrieving API data to be logged",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "parameters" : [ {
          "in" : "body",
          "name" : "Payload",
          "description" : "Request Body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/APIList"
          }
        }, {
          "name" : "context",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "logLevel",
          "in" : "query",
          "required" : false,
          "type" : "string",
          "enum" : [ "all", "headers", "body" ]
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          },
          "404" : {
            "description" : "Not Found.\nRequest API resource or external store Ids not found.\n",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal server error while configuring API to be logged",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "delete" : {
        "parameters" : [ {
          "name" : "context",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : ""
          },
          "404" : {
            "description" : "Not Found.\nRequest API resource or external store Ids not found.\n",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "IInternal server error while deleting API details",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "APIList" : {
      "properties" : {
        "apis" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/API"
          }
        }
      },
      "title" : "apis"
    },
    "API" : {
      "properties" : {
        "context" : {
          "type" : "string",
          "example" : "pizashack/v1.0.0"
        },
        "logLevel" : {
          "type" : "string",
          "example" : "all"
        }
      },
      "title" : "API Info object with log details"
    },
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int64"
        },
        "message" : {
          "type" : "string",
          "description" : "Error message."
        },
        "description" : {
          "type" : "string",
          "description" : "A detail description about the error message.\n"
        },
        "moreInfo" : {
          "type" : "string",
          "description" : "Preferably an url with more details about the error.\n"
        },
        "error" : {
          "type" : "array",
          "description" : "If there are more than one error list them out.\nFor example, list out validation errors by each field.\n",
          "items" : {
            "$ref" : "#/definitions/ErrorListItem"
          }
        }
      },
      "title" : "Error object returned with 4XX HTTP Status"
    },
    "ErrorListItem" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string",
          "description" : "Description about individual errors occurred\n"
        },
        "description" : {
          "type" : "string",
          "description" : "A detail description about the error message."
        }
      },
      "title" : "Description of individual errors that may have occurred during a request."
    }
  }
}